Refactor to cachedRequest
