+++ src/pages/api/releases.js
@@ -1,4 +1,4 @@
-import cachedFetch from "utils/proxy/cached-fetch";
+import { cachedRequest } from "utils/proxy/http";
 import createLogger from "utils/logger";
 
 const logger = createLogger("releases");
@@ -6,7 +6,7 @@ const logger = createLogger("releases");
 export default async function handler(req, res) {
   const releasesURL = "https://api.github.com/repos/gethomepage/homepage/releases";
   try {
-    return res.send(await cachedFetch(releasesURL, 5));
+    return res.send(await cachedRequest(releasesURL, 5));
   } catch (e) {
     logger.error(`Error checking GitHub releases: ${e}`);
     return res.send([]);
